- Update GETTING_STARTED with dev container instructions - Remove detailed instructions from GETTING_STARTED because they are located in the Wiki now - Add POSIX build to init script - Add Git-Graph plugin to VS Code configuration - Add dependencies for POSIX GUI - Add nRF Command Line Tools - Replace static path in default build configurations with dynamic path - Add .bashrc for container initialization - Add Doxygen Documentation Generator plugin - Add remote GDB interface

+++ .devcontainer/init.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-# This script must be run from the project root!
 
 JLINK_VERSION=JLink_Linux_V794b_x86_64
 TEMP_DIR="/tmp/downloads"
@@ -17,20 +16,15 @@ apt update
 apt -y install pkg-config libsdl2-dev:i386
 export PKG_CONFIG_PATH=/usr/lib/i386-linux-gnu/pkgconfig
 
-echo -e "${GREEN}Install nRF Command Line Tools${ENDCOLOR}"
-NRF_CMD_LINE_DEB=${TEMP_DIR}/JLink_Linux_x86_64.deb
-wget https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/desktop-software/nrf-command-line-tools/sw/versions-10-x-x/10-23-2/nrf-command-line-tools_10.23.2_amd64.deb -O ${NRF_CMD_LINE_DEB}
-dpkg -i ${NRF_CMD_LINE_DEB}
-
-echo -e "${GREEN}Enable udev daemon${ENDCOLOR}"
-apt -y install udev
-/lib/systemd/systemd-udevd --daemon
-
-echo -e "${GREEN}Install J-Link driver${ENDCOLOR}"
-echo -e "${RED}Please note: You automatically accept the Segger license terms when you run this installation!${ENDCOLOR}"
-JLINK_DEB=${TEMP_DIR}/JLink_Linux_x86_64.deb
-curl -s -X POST https://www.segger.com/downloads/jlink/${JLINK_VERSION}.deb H "Content-Type: application/x-www-form-urlencoded" -d "accept_license_agreement=accepted" >$JLINK_DEB
-dpkg -i $JLINK_DEB
+if [ ! -f "`which nrfjprog`" ]; then
+	echo -e "${GREEN}Install nRF Command Line Tools${ENDCOLOR}"
+	NRF_CMD_LINE_DEB=${TEMP_DIR}/JLink_Linux_x86_64.deb
+	wget https://nsscprodmedia.blob.core.windows.net/prod/software-and-other-downloads/desktop-software/nrf-command-line-tools/sw/versions-10-x-x/10-23-2/nrf-command-line-tools_10.23.2_amd64.deb -O ${NRF_CMD_LINE_DEB}
+	dpkg -i ${NRF_CMD_LINE_DEB}
+else
+	echo -e "${GREEN}nRF Command Line Tools already installed!${ENDCOLOR}"
+	nrfjprog --version 
+fi
 
 echo -e "${GREEN}Initialize project${ENDCOLOR}"
 cd ${CURRENT_DIR}
@@ -41,8 +35,10 @@ west init -l .
 west update
 
 echo -e "${GREEN}Add build configurations${ENDCOLOR}"
-west build --build-dir /workspaces/ZSWatch/app/debug_rev4 /workspaces/ZSWatch/app --pristine --board zswatch_nrf5340_cpuapp@4 --no-sysbuild -- -DNCS_TOOLCHAIN_VERSION=NONE -DBOARD_ROOT=/workspaces/ZSWatch/app -DCONF_FILE=/workspaces/ZSWatch/app/prj.conf -DEXTRA_CONF_FILE=/workspaces/ZSWatch/app/boards/debug.conf
-west build --build-dir /workspaces/ZSWatch/app/release_rev4 /workspaces/ZSWatch/app --pristine --board zswatch_nrf5340_cpuapp@4 --no-sysbuild -- -DNCS_TOOLCHAIN_VERSION=NONE -DBOARD_ROOT=/workspaces/ZSWatch/app -DCONF_FILE=/workspaces/ZSWatch/app/prj.conf -DEXTRA_CONF_FILE=/workspaces/ZSWatch/app/boards/release.conf
+west build --build-dir ${PWD}/debug_rev4 ${PWD}/app --pristine --board zswatch_nrf5340_cpuapp@4 --no-sysbuild -- -DNCS_TOOLCHAIN_VERSION=NONE -DBOARD_ROOT=/workspaces/ZSWatch/app -DCONF_FILE=/workspaces/ZSWatch/app/prj.conf -DEXTRA_CONF_FILE=${PWD}/app/boards/debug.conf
+west build --build-dir ${PWD}/release_rev4 ${PWD}/app --pristine --board zswatch_nrf5340_cpuapp@4 --no-sysbuild -- -DNCS_TOOLCHAIN_VERSION=NONE -DBOARD_ROOT=/workspaces/ZSWatch/app -DCONF_FILE=/workspaces/ZSWatch/app/prj.conf -DEXTRA_CONF_FILE=${PWD}/app/boards/release.conf
+west build --build-dir ${PWD}/native_posix ${PWD}/app --pristine --board native_posix -- -DOVERLAY_CONFIG=boards/native_posix.conf
 
 echo -e "${GREEN}Clean up${ENDCOLOR}"
-rm -rf ${PWD}/${TEMP_DIR}
+touch /tmp/initialized
+rm -rf ${TEMP_DIR}
